[ODA-254] Minor fixes in I2C

+++ oda-core/commons/src/main/java/es/amplia/oda/core/commons/i2c/I2CDevice.java
@@ -26,7 +26,7 @@ public interface I2CDevice {
 	 * Read an unsigned integer from the device (on the file of the controller, in the address and register specified).
 	 * @return long representing the actual value read from the device.
 	 */
-	long readUInt();
+	double readUInt();
 
 	/**
 	 * Read the specified amount of bytes and return it in a byteBuffer.
